better names and cut old flat map out
