+++ src/flat_hash_map.c
@@ -1,53 +1,12 @@
-/** This implementation is a Robin Hood hash table with shift delete. It caches
-the hash values for efficient distance calculations, resizing, and faster
-comparison before being forced to call user comparison callback. It also
-implements backshift deletions. At first I thought this was not a good starter
-implementation but it actually has some nice features for its required use case
-
-    - Robin Hood hash tables do not use tombstones. They suffer from primary
-      clustering under linear probing, but deletions do allow us to heal the
-      table in some cases. We also try to mitigate primary clustering by
-      choosing prime number table sizes. This is very beneficial when the hash
-      table is not initialized with allocation permission. Long term use of a
-      fixed size table is a use case we must support.
-    - Strict-aliasing is an issue for more complex table schemes. Here we
-      require an intrusive element in the user type meaning they can pass us
-      a block of their type for us to manage as a hash table. If we were to
-      attempt to implement a small metadata array at the start and the user
-      type after that--with no intrusive elements--we would need to likely
-      expose some sort of new type to the user for initialization to help
-      them avoid passing us just some generic char buffer they have on the
-      data segment, stack, or heap; that would be a strict-aliasing violation.
-    - Tailoring callback resistance. Right now, the table caches the full hash
-      for the user because it is needed to accurately calculate distances to
-      the home slot for the Robin Hood swapping purposes. This also helps
-      prevent calling the user comparison callback which is nice; in this table
-      if the user hash is good, they could hash strings and still operate at
-      high efficiency with likely few collisions in the find op. This is a
-      higher space overhead choice, however.
-
-Those are some considerations I can think of now. When attempting to improve
-this table we need to consider these points.
-
-    - The hash table must support a non-allocating mode without the ability to
-      resize to correct from any type of tombstone overload; an in-place
-      rehash would be the only option (not sure how to implement this).
-    - It must work with the user choosing where the memory comes from for the
-      table. This is the other side of non-allocating mode. The user should
-      be able to give us memory coming from the stack, heap, or data segment.
-    - No strict-aliasing violations. I want this library to be useful if the
-      user wants to write correct C code meaning it cannot violate this.
-
-Overall, this means any new design needs slightly more consideration than I
-first thought. However, improvements can still be made. */
+#include <assert.h>
+#include <emmintrin.h>
+#include <limits.h>
+#include <stdalign.h>
 #include <stddef.h>
-#include <stdint.h>
 #include <string.h>
 
-#include "buffer.h"
 #include "flat_hash_map.h"
 #include "impl/impl_flat_hash_map.h"
-#include "impl/impl_types.h"
 #include "types.h"
 
 #if defined(__GNUC__) || defined(__clang__)
@@ -58,123 +17,118 @@ first thought. However, improvements can still be made. */
 #    define likely(expr) expr
 #endif
 
-/* We are fairly close to the maximum 64 bit address space size by the last
-   prime so we will stop there as max size if we ever get there. Not likely.
-   The last prime is closer to the second to last because doubling breaks down
-   at the end. */
-static size_t const primes[] = {
-    11U,
-    37U,
-    79U,
-    163U,
-    331U,
-    691U,
-    1439U,
-    2999U,
-    6079U,
-    12263U,
-    25717U,
-    53611U,
-    111697U,
-    232499U,
-    483611U,
-    1005761U,
-    2091191U,
-    4347799U,
-    9039167U,
-    18792019U,
-    39067739U,
-    81219493U,
-    168849973U,
-    351027263U,
-    729760741U,
-    1517120861U,
-    3153985903U,
-    6556911073U,
-    13631348041U,
-    28338593999U,
-    58913902231U,
-    122477772247U,
-    254622492793U,
-    529341875947U,
-    1100463743389U,
-    2287785087839U,
-    4756140888377U,
-    9887675318611U,
-    20555766849589U,
-    42733962953639U,
-    88840839803449U,
-    184693725350723U,
-    383964990193007U,
-    798235638020831U,
-    1659474561692683U,
-    3449928429320413U,
-    7172153428667531U,
-    14910391869919981U,
-    30997633824443711U,
-    64441854452711651U,
-    133969987155270011U,
-    278513981492471381U,
-    579010564484755961U,
-    1203721378684243091U,
-    2502450294306576181U,
-    5202414434410211531U,
-    10815445968671840317U,
-    17617221824571301183U,
+/* Can we vectorize? Single field unions are guaranteed by the standard to have
+no padding even though some claim single field structs get the same benefit,
+trivially. However, not guaranteed by the standard so use unions instead */
+#if defined(__x86_64) && defined(__SSE2__) && !defined(CCC_FHM_PORTABLE)
+#    include <immintrin.h>
+
+typedef struct
+{
+    __m128i v;
+} group;
+
+typedef struct
+{
+    uint16_t v;
+} index_mask;
+
+enum : typeof((index_mask){}.v)
+{
+    INDEX_MASK_MSB = 0x8000,
 };
+
+#else
+#    include <stdint.h>
+typedef struct
+{
+    uint64_t v;
+} group;
+
+typedef struct
+{
+    typeof((group){}.v) v;
+} index_mask;
+
+enum : typeof((index_mask){}.v)
+{
+    INDEX_MASK_MSBYTE = 0xFF00000000000000,
+};
+
+enum : typeof((group){}.v)
+{
+    GROUP_LSB = 0x101010101010101,
+    GROUP_DELETED = 0x8080808080808080,
+};
+
+#endif /* defined(__x86_64) && defined(__SSE2__) &&                            \
+          !defined(CCC_FHM_GENERIC)*/
 
 enum : size_t
 {
-    PRIMES_SIZE = ((sizeof(primes) / sizeof(primes[0]))),
-    LAST_SWAP_SLOT = 1,
-    NUM_SWAP_SLOTS = 2,
+    GROUP_BYTES = sizeof(group),
+    GROUP_BITS = GROUP_BYTES * CHAR_BIT,
+    SIZE_T_MSB = 0x1000000000000000,
 };
 
-/* Some claim that Robin Hood tables can support a much higher load factor. I
-would not be so sure. The primary clustering effect in these types of tables can
-quickly rise. Try to mitigate with a lower load factor and prime table sizes.
-Measure. */
-static double const load_factor = 0.8;
+enum : uint8_t
+{
+    TAG_MSB = 0x80,
+    TAG_LSB = 0x1,
+    LOWER_7_BITS_MASK = 0x7F,
+    TAG_BITS = sizeof(ccc_fhm_tag) * CHAR_BIT,
+};
 
-/*=========================   Prototypes   ==================================*/
+struct triangular_seq
+{
+    size_t i;
+    size_t stride;
+};
 
-static void erase(struct ccc_fhmap_ *, void *);
-static void swap(char tmp[const], void *, void *, size_t);
-static void *struct_base(struct ccc_fhmap_ const *,
-                         struct ccc_fhmap_elem_ const *);
-static struct ccc_ent_ entry(struct ccc_fhmap_ *, void const *key,
-                             uint64_t hash);
+/*===========================   Prototypes   ================================*/
+
 static struct ccc_fhash_entry_ container_entry(struct ccc_fhmap_ *h,
                                                void const *key);
-static ccc_tribool valid_distance_from_home(struct ccc_fhmap_ const *,
-                                            void const *slot);
-static size_t to_i(size_t capacity, uint64_t hash);
-static size_t increment(size_t capacity, size_t i);
-static size_t decrement(size_t capacity, size_t i);
-static size_t distance(size_t capacity, size_t i, size_t j);
-static void *key_in_slot(struct ccc_fhmap_ const *h, void const *slot);
-static struct ccc_fhmap_elem_ *elem_in_slot(struct ccc_fhmap_ const *h,
-                                            void const *slot);
-static ccc_result maybe_resize(struct ccc_fhmap_ *h);
-static struct ccc_ent_ find(struct ccc_fhmap_ const *h, void const *key,
-                            uint64_t hash);
-static void insert(struct ccc_fhmap_ *h, void const *e, uint64_t hash,
-                   size_t i);
-static uint64_t *hash_at(struct ccc_fhmap_ const *h, size_t i);
-static uint64_t filter(struct ccc_fhmap_ const *h, void const *key);
-static size_t next_prime(size_t n);
+static struct ccc_handl_ handle(struct ccc_fhmap_ *h, void const *key,
+                                uint64_t hash);
+static struct ccc_handl_ find_key_or_slot(struct ccc_fhmap_ const *h,
+                                          void const *key, uint64_t hash);
+static ccc_ucount find_key(struct ccc_fhmap_ const *h, void const *key,
+                           uint64_t hash);
+static size_t find_known_insert_slot(struct ccc_fhmap_ const *h, uint64_t hash);
+static ccc_result maybe_rehash(struct ccc_fhmap_ *h, size_t to_add);
+static void insert(struct ccc_fhmap_ *h, void const *key_val_type,
+                   ccc_fhm_tag m, size_t i);
+static void erase(struct ccc_fhmap_ *h, size_t i);
+static void rehash_in_place(struct ccc_fhmap_ *h);
+static ccc_result rehash_resize(struct ccc_fhmap_ *h, size_t to_add);
+static void *key_at(struct ccc_fhmap_ const *h, size_t i);
+static void *data_at(struct ccc_fhmap_ const *h, size_t i);
+static void swap(char tmp[const], void *a, void *b, size_t ab_size);
 
-/*=========================   Interface    ==================================*/
+static void set_tag(struct ccc_fhmap_ *h, ccc_fhm_tag m, size_t i);
+static ccc_tribool is_index_on(index_mask m);
+static size_t lowest_on_index(index_mask m);
+static size_t trailing_zeros(index_mask m);
+static size_t leading_zeros(index_mask m);
+static size_t next_index(index_mask *m);
+static ccc_tribool is_full(ccc_fhm_tag m);
+static ccc_tribool is_constant(ccc_fhm_tag m);
+static ccc_tribool is_empty_constant(ccc_fhm_tag m);
+static ccc_fhm_tag to_tag(uint64_t hash);
+static group load_group(ccc_fhm_tag const *src);
+static void store_group(ccc_fhm_tag *dst, group src);
+static index_mask match_tag(group g, ccc_fhm_tag m);
+static index_mask match_empty(group g);
+static index_mask match_empty_or_deleted(group g);
+static index_mask match_full(group g);
+static group make_deleted_empty_and_full_deleted(group g);
+static unsigned countr_0(index_mask m);
+static unsigned countl_0(index_mask m);
+static unsigned countl_0_size_t(size_t n);
+static size_t next_power_of_two(size_t n);
 
-ccc_tribool
-ccc_fhm_is_empty(ccc_flat_hash_map const *const h)
-{
-    if (unlikely(!h))
-    {
-        return CCC_TRIBOOL_ERROR;
-    }
-    return !ccc_fhm_size(h).count;
-}
+/*===========================    Interface   ================================*/
 
 ccc_tribool
 ccc_fhm_contains(ccc_flat_hash_map *const h, void const *const key)
@@ -183,23 +137,15 @@ ccc_fhm_contains(ccc_flat_hash_map *const h, void const *const key)
     {
         return CCC_TRIBOOL_ERROR;
     }
-    return (entry(h, key, filter(h, key)).stats_ & CCC_ENTRY_OCCUPIED) != 0;
-}
-
-ccc_ucount
-ccc_fhm_size(ccc_flat_hash_map const *const h)
-{
-    if (unlikely(!h))
+    if (unlikely(!h->init_ || !h->sz_))
     {
-        return (ccc_ucount){.error = CCC_RESULT_ARG_ERROR};
+        return CCC_FALSE;
     }
-    ccc_ucount sz = ccc_buf_size(&h->buf_);
-    if (sz.error || !sz.count)
-    {
-        return sz;
-    }
-    sz.count -= NUM_SWAP_SLOTS;
-    return sz;
+    return CCC_RESULT_OK
+           == find_key(
+                  h, key,
+                  h->hash_fn_((ccc_user_key){.user_key = key, .aux = h->aux_}))
+                  .error;
 }
 
 ccc_fhmap_entry
@@ -207,51 +153,31 @@ ccc_fhm_entry(ccc_flat_hash_map *const h, void const *const key)
 {
     if (unlikely(!h || !key))
     {
-        return (ccc_fhmap_entry){{.entry_ = {.stats_ = CCC_ENTRY_ARG_ERROR}}};
+        return (ccc_fhmap_entry){{.handle_ = {.stats_ = CCC_ENTRY_ARG_ERROR}}};
     }
     return (ccc_fhmap_entry){container_entry(h, key)};
 }
 
 void *
-ccc_fhm_insert_entry(ccc_fhmap_entry const *const e, ccc_fhmap_elem *const elem)
+ccc_fhm_insert_entry(ccc_fhmap_entry const *const e, void const *key_val_type)
 {
-    if (unlikely(!e || !elem))
-    {
-        return NULL;
-    }
-    void *const user_struct = struct_base(e->impl_.h_, elem);
-    if (e->impl_.entry_.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        elem->hash_ = e->impl_.hash_;
-        if (e->impl_.entry_.e_ != user_struct && e->impl_.entry_.e_)
-        {
-            (void)memcpy(e->impl_.entry_.e_, user_struct,
-                         e->impl_.h_->buf_.elem_sz_);
-        }
-        return e->impl_.entry_.e_;
-    }
-    if (e->impl_.entry_.stats_ & CCC_ENTRY_INSERT_ERROR)
-    {
-        return NULL;
-    }
-    insert(e->impl_.h_, user_struct, e->impl_.hash_,
-           ccc_buf_i(&e->impl_.h_->buf_, e->impl_.entry_.e_).count);
-    return e->impl_.entry_.e_;
-}
 
-void *
-ccc_fhm_get_key_val(ccc_flat_hash_map *const h, void const *const key)
-{
-    if (unlikely(!h || !key || !h->buf_.capacity_))
+    if (unlikely(!e || !key_val_type))
     {
         return NULL;
     }
-    struct ccc_ent_ e = find(h, key, filter(h, key));
-    if (e.stats_ & CCC_ENTRY_OCCUPIED)
+    if (e->impl_.handle_.stats_ & CCC_ENTRY_OCCUPIED)
     {
-        return e.e_;
+        void *const slot = data_at(e->impl_.h_, e->impl_.handle_.i_);
+        (void)memcpy(slot, key_val_type, e->impl_.h_->elem_sz_);
+        return slot;
     }
-    return NULL;
+    if (e->impl_.handle_.stats_ & CCC_ENTRY_INSERT_ERROR)
+    {
+        return NULL;
+    }
+    insert(e->impl_.h_, key_val_type, e->impl_.tag_, e->impl_.handle_.i_);
+    return data_at(e->impl_.h_, e->impl_.handle_.i_);
 }
 
 ccc_entry
@@ -261,744 +187,401 @@ ccc_fhm_remove_entry(ccc_fhmap_entry const *const e)
     {
         return (ccc_entry){{.stats_ = CCC_ENTRY_ARG_ERROR}};
     }
-    if (e->impl_.entry_.stats_ != CCC_ENTRY_OCCUPIED)
+    if (e->impl_.handle_.stats_ != CCC_ENTRY_OCCUPIED)
     {
         return (ccc_entry){{.stats_ = CCC_ENTRY_VACANT}};
     }
-    erase(e->impl_.h_, e->impl_.entry_.e_);
-    return (ccc_entry){{.stats_ = CCC_ENTRY_OCCUPIED}};
+    erase(e->impl_.h_, e->impl_.handle_.i_);
+    return (ccc_entry){{.e_ = data_at(e->impl_.h_, e->impl_.handle_.i_),
+                        .stats_ = CCC_ENTRY_OCCUPIED}};
 }
 
-ccc_fhmap_entry *
-ccc_fhm_and_modify(ccc_fhmap_entry *const e, ccc_update_fn *const fn)
-{
-    if (e && fn && e->impl_.entry_.stats_ == CCC_ENTRY_OCCUPIED
-        && e->impl_.entry_.e_)
-    {
-        fn((ccc_user_type){e->impl_.entry_.e_, NULL});
-    }
-    return e;
-}
-
-ccc_fhmap_entry *
-ccc_fhm_and_modify_aux(ccc_fhmap_entry *const e, ccc_update_fn *const fn,
-                       void *const aux)
-{
-    if (fn && e && e->impl_.entry_.stats_ == CCC_ENTRY_OCCUPIED
-        && e->impl_.entry_.e_)
-    {
-        fn((ccc_user_type){e->impl_.entry_.e_, aux});
-    }
-    return e;
-}
-
-ccc_entry
-ccc_fhm_swap_entry(ccc_flat_hash_map *const h, ccc_fhmap_elem *const out_handle)
-{
-    if (unlikely(!h || !out_handle))
-    {
-        return (ccc_entry){{.stats_ = CCC_ENTRY_ARG_ERROR}};
-    }
-    void *const user_return = struct_base(h, out_handle);
-    void *const key = key_in_slot(h, user_return);
-    size_t const user_struct_size = h->buf_.elem_sz_;
-    struct ccc_fhash_entry_ ent = container_entry(h, key);
-    if (ent.entry_.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        out_handle->hash_ = ent.hash_;
-        void *const tmp = ccc_buf_at(&h->buf_, 0);
-        swap(tmp, ent.entry_.e_, user_return, user_struct_size);
-        *hash_at(h, 0) = CCC_FHM_EMPTY;
-        return (ccc_entry){{.e_ = user_return, .stats_ = CCC_ENTRY_OCCUPIED}};
-    }
-    if (ent.entry_.stats_ & CCC_ENTRY_INSERT_ERROR || !ent.entry_.e_)
-    {
-        return (ccc_entry){{.stats_ = CCC_ENTRY_INSERT_ERROR}};
-    }
-    insert(h, user_return, ent.hash_, ccc_buf_i(&h->buf_, ent.entry_.e_).count);
-    return (ccc_entry){{.stats_ = CCC_ENTRY_VACANT}};
-}
-
-ccc_entry
-ccc_fhm_try_insert(ccc_flat_hash_map *const h,
-                   ccc_fhmap_elem *const key_val_handle)
-{
-    if (unlikely(!h || !key_val_handle))
-    {
-        return (ccc_entry){{.stats_ = CCC_ENTRY_ARG_ERROR}};
-    }
-    void *const user_base = struct_base(h, key_val_handle);
-    struct ccc_fhash_entry_ ent = container_entry(h, key_in_slot(h, user_base));
-    if (ent.entry_.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        return (ccc_entry){{.e_ = ent.entry_.e_, .stats_ = CCC_ENTRY_OCCUPIED}};
-    }
-    if (ent.entry_.stats_ & CCC_ENTRY_INSERT_ERROR || !ent.entry_.e_)
-    {
-        return (ccc_entry){{.e_ = NULL, .stats_ = CCC_ENTRY_INSERT_ERROR}};
-    }
-    insert(h, user_base, ent.hash_, ccc_buf_i(&h->buf_, ent.entry_.e_).count);
-    return (ccc_entry){{.e_ = ent.entry_.e_, .stats_ = CCC_ENTRY_VACANT}};
-}
-
-ccc_entry
-ccc_fhm_insert_or_assign(ccc_flat_hash_map *const h,
-                         ccc_fhmap_elem *const key_val_handle)
-{
-    if (unlikely(!h || !key_val_handle))
-    {
-        return (ccc_entry){{.stats_ = CCC_ENTRY_ARG_ERROR}};
-    }
-    void *const user_base = struct_base(h, key_val_handle);
-    struct ccc_fhash_entry_ ent = container_entry(h, key_in_slot(h, user_base));
-    if (ent.entry_.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        key_val_handle->hash_ = ent.hash_;
-        if (ent.entry_.e_ != user_base)
-        {
-            (void)memcpy(ent.entry_.e_, user_base, h->buf_.elem_sz_);
-        }
-        return (ccc_entry){{.e_ = ent.entry_.e_, .stats_ = CCC_ENTRY_OCCUPIED}};
-    }
-    if (ent.entry_.stats_ & CCC_ENTRY_INSERT_ERROR || !ent.entry_.e_)
-    {
-        return (ccc_entry){{.e_ = NULL, .stats_ = CCC_ENTRY_INSERT_ERROR}};
-    }
-    insert(h, user_base, ent.hash_, ccc_buf_i(&h->buf_, ent.entry_.e_).count);
-    return (ccc_entry){{.e_ = ent.entry_.e_, .stats_ = CCC_ENTRY_VACANT}};
-}
-
-ccc_entry
-ccc_fhm_remove(ccc_flat_hash_map *const h, ccc_fhmap_elem *const out_handle)
-{
-    if (unlikely(!h || !out_handle))
-    {
-        return (ccc_entry){{.stats_ = CCC_ENTRY_ARG_ERROR}};
-    }
-    void *const ret = struct_base(h, out_handle);
-    void *const key = key_in_slot(h, ret);
-    struct ccc_ent_ const ent = find(h, key, filter(h, key));
-    if (ent.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        if (ret != ent.e_)
-        {
-            (void)memcpy(ret, ent.e_, h->buf_.elem_sz_);
-        }
-        erase(h, ent.e_);
-        return (ccc_entry){{.e_ = ret, .stats_ = CCC_ENTRY_OCCUPIED}};
-    }
-    return (ccc_entry){{.e_ = NULL, .stats_ = CCC_ENTRY_VACANT}};
-}
-
-void *
-ccc_fhm_or_insert(ccc_fhmap_entry const *const e, ccc_fhmap_elem *const elem)
-{
-    if (unlikely(!e || !elem))
-    {
-        return NULL;
-    }
-    if (e->impl_.entry_.stats_ & CCC_ENTRY_OCCUPIED)
-    {
-        return e->impl_.entry_.e_;
-    }
-    if (e->impl_.entry_.stats_ & CCC_ENTRY_INSERT_ERROR || !e->impl_.entry_.e_)
-    {
-        return NULL;
-    }
-    void *user_struct = struct_base(e->impl_.h_, elem);
-    elem->hash_ = e->impl_.hash_;
-    insert(e->impl_.h_, user_struct, elem->hash_,
-           ccc_buf_i(&e->impl_.h_->buf_, e->impl_.entry_.e_).count);
-    return e->impl_.entry_.e_;
-}
-
-void *
-ccc_fhm_unwrap(ccc_fhmap_entry const *const e)
-{
-    if (unlikely(!e) || !(e->impl_.entry_.stats_ & CCC_ENTRY_OCCUPIED))
-    {
-        return NULL;
-    }
-    return e->impl_.entry_.e_;
-}
-
-ccc_tribool
-ccc_fhm_occupied(ccc_fhmap_entry const *const e)
-{
-    if (unlikely(!e))
-    {
-        return CCC_TRIBOOL_ERROR;
-    }
-    return (e->impl_.entry_.stats_ & CCC_ENTRY_OCCUPIED) != 0;
-}
-
-ccc_tribool
-ccc_fhm_insert_error(ccc_fhmap_entry const *const e)
-{
-    if (unlikely(!e))
-    {
-        return CCC_TRIBOOL_ERROR;
-    }
-    return (e->impl_.entry_.stats_ & CCC_ENTRY_INSERT_ERROR) != 0;
-}
-
-ccc_entry_status
-ccc_fhm_entry_status(ccc_fhmap_entry const *const e)
-{
-    if (unlikely(!e))
-    {
-        return CCC_ENTRY_ARG_ERROR;
-    }
-    return e->impl_.entry_.stats_;
-}
-
-void *
-ccc_fhm_begin(ccc_flat_hash_map const *const h)
-{
-    if (unlikely(!h || ccc_buf_is_empty(&h->buf_)))
-    {
-        return NULL;
-    }
-    void const *iter = ccc_buf_begin(&h->buf_);
-    for (; iter != ccc_buf_capacity_end(&h->buf_)
-           && elem_in_slot(h, iter)->hash_ == CCC_FHM_EMPTY;
-         iter = ccc_buf_next(&h->buf_, iter))
-    {}
-    return (void *)(iter == ccc_buf_capacity_end(&h->buf_) ? NULL : iter);
-}
-
-void *
-ccc_fhm_next(ccc_flat_hash_map const *const h, ccc_fhmap_elem const *const iter)
-{
-    if (unlikely(!h))
-    {
-        return NULL;
-    }
-    void const *i = struct_base(h, iter);
-    for (i = ccc_buf_next(&h->buf_, i);
-         i != ccc_buf_capacity_end(&h->buf_)
-         && elem_in_slot(h, i)->hash_ == CCC_FHM_EMPTY;
-         i = ccc_buf_next(&h->buf_, i))
-    {}
-    return (void *)(i == ccc_buf_capacity_end(&h->buf_) ? NULL : i);
-}
-
-void *
-ccc_fhm_end(ccc_flat_hash_map const *const)
-{
-    return NULL;
-}
-
-ccc_ucount
-ccc_fhm_next_prime(size_t const n)
-{
-    size_t const next = next_prime(n);
-    if (!next)
-    {
-        return (ccc_ucount){.error = CCC_RESULT_FAIL};
-    }
-    return (ccc_ucount){.count = next};
-}
-
-ccc_result
-ccc_fhm_copy(ccc_flat_hash_map *const dst, ccc_flat_hash_map const *const src,
-             ccc_alloc_fn *const fn)
-{
-    if (!dst || !src || src == dst
-        || (dst->buf_.capacity_ < src->buf_.capacity_ && !fn))
-    {
-        return CCC_RESULT_ARG_ERROR;
-    }
-    /* Copy everything so we don't worry about staying in sync with future
-       changes to buf container. But we have to give back original destination
-       memory in case it has already been allocated. Alloc will remain the
-       same as in dst initialization because that controls permission. */
-    void *const dst_mem = dst->buf_.mem_;
-    size_t const dst_cap = dst->buf_.capacity_;
-    ccc_alloc_fn *const dst_alloc = dst->buf_.alloc_;
-    *dst = *src;
-    dst->buf_.mem_ = dst_mem;
-    dst->buf_.capacity_ = dst_cap;
-    dst->buf_.alloc_ = dst_alloc;
-    if (!src->buf_.capacity_)
-    {
-        return CCC_RESULT_OK;
-    }
-    if (dst->buf_.capacity_ < src->buf_.capacity_)
-    {
-        ccc_result resize_res
-            = ccc_buf_alloc(&dst->buf_, src->buf_.capacity_, fn);
-        if (resize_res != CCC_RESULT_OK)
-        {
-            return resize_res;
-        }
-        dst->buf_.capacity_ = src->buf_.capacity_;
-    }
-    if (!dst->buf_.mem_ || !src->buf_.mem_)
-    {
-        return CCC_RESULT_ARG_ERROR;
-    }
-    if (dst->buf_.capacity_ == src->buf_.capacity_)
-    {
-        (void)memcpy(dst->buf_.mem_, src->buf_.mem_,
-                     src->buf_.capacity_ * src->buf_.elem_sz_);
-        return CCC_RESULT_OK;
-    }
-    dst->buf_.sz_ = NUM_SWAP_SLOTS;
-    for (void *slot = ccc_buf_begin(&src->buf_);
-         slot != ccc_buf_capacity_end(&src->buf_);
-         slot = ccc_buf_next(&src->buf_, slot))
-    {
-        struct ccc_fhmap_elem_ const *const e = elem_in_slot(src, slot);
-        if (e->hash_ != CCC_FHM_EMPTY)
-        {
-            struct ccc_ent_ const new_ent
-                = find(dst, key_in_slot(src, slot), e->hash_);
-            insert(dst, slot, e->hash_,
-                   ccc_buf_i(&dst->buf_, new_ent.e_).count);
-        }
-    }
-    return CCC_RESULT_OK;
-}
-
-ccc_result
-ccc_fhm_clear(ccc_flat_hash_map *const h, ccc_destructor_fn *const fn)
-{
-    if (unlikely(!h))
-    {
-        return CCC_RESULT_ARG_ERROR;
-    }
-    if (!fn)
-    {
-        h->buf_.sz_ = 0;
-        return CCC_RESULT_OK;
-    }
-    for (void *slot = ccc_buf_begin(&h->buf_);
-         slot != ccc_buf_capacity_end(&h->buf_);
-         slot = ccc_buf_next(&h->buf_, slot))
-    {
-        if (elem_in_slot(h, slot)->hash_ != CCC_FHM_EMPTY)
-        {
-            fn((ccc_user_type){.user_type = slot, .aux = h->buf_.aux_});
-        }
-    }
-    return CCC_RESULT_OK;
-}
-
-ccc_result
-ccc_fhm_clear_and_free(ccc_flat_hash_map *const h, ccc_destructor_fn *const fn)
-{
-    if (unlikely(!h))
-    {
-        return CCC_RESULT_ARG_ERROR;
-    }
-    if (!fn)
-    {
-        h->buf_.sz_ = 0;
-        return ccc_buf_alloc(&h->buf_, 0, h->buf_.alloc_);
-    }
-    for (void *slot = ccc_buf_begin(&h->buf_);
-         slot != ccc_buf_capacity_end(&h->buf_);
-         slot = ccc_buf_next(&h->buf_, slot))
-    {
-        if (elem_in_slot(h, slot)->hash_ != CCC_FHM_EMPTY)
-        {
-            fn((ccc_user_type){.user_type = slot, .aux = h->buf_.aux_});
-        }
-    }
-    return ccc_buf_alloc(&h->buf_, 0, h->buf_.alloc_);
-}
-
-ccc_ucount
-ccc_fhm_capacity(ccc_flat_hash_map const *const h)
-{
-    if (unlikely(!h))
-    {
-        return (ccc_ucount){.error = CCC_RESULT_ARG_ERROR};
-    }
-    return ccc_buf_capacity(&h->buf_);
-}
-
-void *
-ccc_fhm_data(ccc_flat_hash_map const *const h)
-{
-    return h ? ccc_buf_begin(&h->buf_) : NULL;
-}
-
-ccc_tribool
-ccc_fhm_validate(ccc_flat_hash_map const *const h)
-{
-    if (!h || !h->eq_fn_ || !h->hash_fn_)
-    {
-        return CCC_TRIBOOL_ERROR;
-    }
-    size_t empties = 0;
-    size_t occupied = 0;
-    for (void const *i = ccc_buf_begin(&h->buf_);
-         i != ccc_buf_capacity_end(&h->buf_); i = ccc_buf_next(&h->buf_, i))
-    {
-        if (elem_in_slot(h, i)->hash_ == CCC_FHM_EMPTY)
-        {
-            ++empties;
-        }
-        else
-        {
-            ++occupied;
-        }
-        if (elem_in_slot(h, i)->hash_ != CCC_FHM_EMPTY
-            && !valid_distance_from_home(h, i))
-        {
-            return CCC_FALSE;
-        }
-    }
-    return occupied == ccc_fhm_size(h).count
-           && empties == (h->buf_.capacity_ - occupied);
-}
-
-static ccc_tribool
-valid_distance_from_home(struct ccc_fhmap_ const *const h,
-                         void const *const slot)
-{
-    size_t const cap = h->buf_.capacity_;
-    uint64_t const hash = elem_in_slot(h, slot)->hash_;
-    size_t const home = to_i(cap, hash);
-    size_t const end = decrement(cap, home);
-    for (size_t i = ccc_buf_i(&h->buf_, slot).count,
-                distance_to_home = distance(cap, i, to_i(cap, hash));
-         i != end; --distance_to_home, i = decrement(cap, i))
-    {
-        uint64_t const cur_hash = *hash_at(h, i);
-        /* The only reason an element is not home is because it has been
-           moved away to help another element be closer to its home. This
-           would break the purpose of doing that. Upon erase everyone needs
-           to shuffle closer to home. */
-        if (cur_hash == CCC_FHM_EMPTY)
-        {
-            return CCC_FALSE;
-        }
-        /* This shouldn't happen either. The whole point of Robin Hood is
-           taking from the close and giving to the far. If this happens
-           we have made our algorithm greedy not altruistic. */
-        if (distance_to_home > distance(cap, i, to_i(cap, cur_hash)))
-        {
-            return CCC_FALSE;
-        }
-    }
-    return CCC_TRUE;
-}
-
-/*=======================   Private Interface   =============================*/
-
-struct ccc_fhash_entry_
-ccc_impl_fhm_entry(struct ccc_fhmap_ *const h, void const *const key)
-{
-    return container_entry(h, key);
-}
-
-void
-ccc_impl_fhm_insert(struct ccc_fhmap_ *const h, void const *const e,
-                    uint64_t const hash, size_t cur_i)
-{
-    insert(h, e, hash, cur_i);
-}
-
-struct ccc_fhmap_elem_ *
-ccc_impl_fhm_in_slot(struct ccc_fhmap_ const *const h, void const *const slot)
-{
-    return elem_in_slot(h, slot);
-}
-
-void *
-ccc_impl_fhm_key_in_slot(struct ccc_fhmap_ const *const h,
-                         void const *const slot)
-{
-    return key_in_slot(h, slot);
-}
-
-size_t
-ccc_impl_fhm_increment(size_t const capacity, size_t const i)
-{
-    return increment(capacity, i);
-}
-
-uint64_t *
-ccc_impl_fhm_hash_at(struct ccc_fhmap_ const *const h, size_t const i)
-{
-    return hash_at(h, i);
-}
-
-/*=======================     Static Helpers    =============================*/
+/*=========================   Static Internals   ============================*/
 
 static struct ccc_fhash_entry_
 container_entry(struct ccc_fhmap_ *const h, void const *const key)
 {
-    uint64_t const hash = filter(h, key);
+    uint64_t const hash = h->hash_fn_((ccc_user_key){key, h->aux_});
     return (struct ccc_fhash_entry_){
-        .h_ = h,
-        .hash_ = hash,
-        .entry_ = entry(h, key, hash),
+        .h_ = (struct ccc_fhmap_ *)h,
+        .tag_ = to_tag(hash),
+        .handle_ = handle(h, key, hash),
     };
 }
 
-static struct ccc_ent_
-entry(struct ccc_fhmap_ *const h, void const *const key, uint64_t const hash)
+static struct ccc_handl_
+handle(struct ccc_fhmap_ *const h, void const *const key, uint64_t const hash)
 {
     ccc_entry_status upcoming_insertion_error = 0;
-    if (maybe_resize(h) != CCC_RESULT_OK)
+    switch (maybe_rehash(h, 1))
     {
+    case CCC_RESULT_OK:
+        break;
+    case CCC_RESULT_ARG_ERROR:
+        return (struct ccc_handl_){.stats_ = CCC_ENTRY_ARG_ERROR};
+        break;
+    default:
         upcoming_insertion_error = CCC_ENTRY_INSERT_ERROR;
-    }
-    struct ccc_ent_ res = find(h, key, hash);
+        break;
+    };
+    struct ccc_handl_ res = find_key_or_slot(h, key, hash);
     res.stats_ |= upcoming_insertion_error;
     return res;
 }
 
-static struct ccc_ent_
-find(struct ccc_fhmap_ const *const h, void const *const key,
-     uint64_t const hash)
+static void
+insert(struct ccc_fhmap_ *const h, void const *const key_val_type,
+       ccc_fhm_tag const m, size_t const i)
 {
-    size_t const cap = h->buf_.capacity_;
-    /* A few sanity checks. The load factor should be managed a full table is
-       never allowed even under no allocation permission because that could
-       lead to an infinite loop and illustrates a degenerate table anyway. */
-    if (unlikely(!cap))
-    {
-        return (struct ccc_ent_){.e_ = NULL, .stats_ = CCC_ENTRY_VACANT};
-    }
-    if (unlikely(h->buf_.sz_ >= h->buf_.capacity_))
-    {
-        return (struct ccc_ent_){.e_ = NULL, .stats_ = CCC_ENTRY_ARG_ERROR};
-    }
-    size_t i = to_i(cap, hash);
-    size_t dist = 0;
-    do
-    {
-        void *const slot = ccc_buf_at(&h->buf_, i);
-        uint64_t const slot_hash = elem_in_slot(h, slot)->hash_;
-        if (slot_hash == CCC_FHM_EMPTY
-            || dist > distance(cap, i, to_i(cap, slot_hash)))
-        {
-            return (struct ccc_ent_){
-                .e_ = slot, .stats_ = CCC_ENTRY_VACANT | CCC_ENTRY_NO_UNWRAP};
-        }
-        if (hash == slot_hash
-            && h->eq_fn_((ccc_key_cmp){
-                .key_lhs = key, .user_type_rhs = slot, .aux = h->buf_.aux_}))
-        {
-            return (struct ccc_ent_){.e_ = slot, .stats_ = CCC_ENTRY_OCCUPIED};
-        }
-        ++dist;
-        i = increment(cap, i);
-    } while (1);
+    assert(i <= h->mask_);
+    assert((m.v & TAG_MSB) == 0);
+    h->avail_ -= is_empty_constant(h->tag_[i]);
+    ++h->sz_;
+    set_tag(h, m, i);
+    (void)memcpy(data_at(h, i), key_val_type, h->elem_sz_);
 }
 
-/* Assumes that element to be inserted does not already exist in the table.
-   Assumes that the table has room for another insertion. Unexpected results
-   may occur if these assumptions are not accommodated. */
 static void
-insert(struct ccc_fhmap_ *const h, void const *const e, uint64_t const hash,
-       size_t i)
+erase(struct ccc_fhmap_ *const h, size_t const i)
 {
-    size_t const elem_sz = h->buf_.elem_sz_;
-    size_t const cap = h->buf_.capacity_;
-    void *const tmp = ccc_buf_at(&h->buf_, 1);
-    void *const floater = ccc_buf_at(&h->buf_, 0);
-    if (floater != e)
-    {
-        (void)memcpy(floater, e, elem_sz);
-    }
+    assert(i <= h->mask_);
+    size_t const i_before = (i - CCC_FHM_GROUP_SIZE) & h->mask_;
+    index_mask const i_before_empty
+        = match_empty(load_group(&h->tag_[i_before]));
+    index_mask const i_empty = match_empty(load_group(&h->tag_[i]));
+    ccc_fhm_tag const m = leading_zeros(i_before_empty) + leading_zeros(i_empty)
+                                  >= CCC_FHM_GROUP_SIZE
+                              ? (ccc_fhm_tag){CCC_FHM_DELETED}
+                              : (ccc_fhm_tag){CCC_FHM_EMPTY};
+    h->avail_ += (m.v == CCC_FHM_EMPTY);
+    --h->sz_;
+    set_tag(h, m, i);
+}
 
-    /* This function cannot modify e and e may be copied over to new
-       insertion from old table. So should this function invariantly assign
-       starting hash to this slot copy for insertion? I think yes so far. */
-    elem_in_slot(h, floater)->hash_ = hash;
-    size_t dist = distance(cap, i, to_i(cap, hash));
+/* Finds the specified hash or first available slot where the hash could be
+inserted. If the element does not exist and a non-occupied slot is returned
+that slot will have been the first empty or deleted slot encountered in the
+probe sequence. This function assumes an empty slot exists in the table. */
+static struct ccc_handl_
+find_key_or_slot(struct ccc_fhmap_ const *const h, void const *const key,
+                 uint64_t const hash)
+{
+    ccc_fhm_tag const tag = to_tag(hash);
+    size_t const mask = h->mask_;
+    struct triangular_seq seq = {.i = hash & mask, .stride = 0};
+    ccc_ucount empty_or_deleted = {.error = CCC_RESULT_FAIL};
     do
     {
-        void *const slot = ccc_buf_at(&h->buf_, i);
-        uint64_t const slot_hash = elem_in_slot(h, slot)->hash_;
-        if (slot_hash == CCC_FHM_EMPTY)
+        group const g = load_group(&h->tag_[seq.i]);
+        index_mask m = match_tag(g, tag);
+        for (size_t i_match = lowest_on_index(m); i_match != CCC_FHM_GROUP_SIZE;
+             i_match = next_index(&m))
         {
-            if (slot != floater)
+            i_match = (seq.i + i_match) & mask;
+            if (h->eq_fn_((ccc_key_cmp){.key_lhs = key,
+                                        .user_type_rhs = data_at(h, i_match),
+                                        .aux = h->aux_}))
             {
-                (void)memcpy(slot, floater, elem_sz);
+                return (struct ccc_handl_){.i_ = i_match,
+                                           .stats_ = CCC_ENTRY_OCCUPIED};
             }
-            (void)ccc_buf_size_plus(&h->buf_, 1);
-            elem_in_slot(h, floater)->hash_ = CCC_FHM_EMPTY;
-            elem_in_slot(h, tmp)->hash_ = CCC_FHM_EMPTY;
-            return;
         }
-        size_t const slot_dist = distance(cap, i, to_i(cap, slot_hash));
-        if (dist > slot_dist)
+        if (empty_or_deleted.error)
         {
-            swap(tmp, floater, slot, elem_sz);
-            dist = slot_dist;
+            size_t const i_take = lowest_on_index(match_empty_or_deleted(g));
+            if (i_take != CCC_FHM_GROUP_SIZE)
+            {
+                empty_or_deleted.count = (seq.i + i_take) & mask;
+                empty_or_deleted.error = CCC_RESULT_OK;
+            }
         }
-        i = increment(cap, i);
-        ++dist;
+        if (is_index_on(match_empty(g)))
+        {
+            return (struct ccc_handl_){.i_ = empty_or_deleted.count,
+                                       .stats_ = CCC_ENTRY_VACANT};
+        }
+        seq.stride += CCC_FHM_GROUP_SIZE;
+        seq.i += seq.stride;
+        seq.i &= mask;
     } while (1);
 }
 
-/* Backshift deletion is important in for this table because it may not be able
-   to allocate. This prevents the need for tombstones which would hurt table
-   quality quickly if we can't resize. */
-static void
-erase(struct ccc_fhmap_ *const h, void *const e)
+/* Finds key or quits when first empty slot is encountered after a group fails
+to match. This function is better when a simple lookup is needed as a few
+branches and loads of groups are omitted compared to the search with intention
+to insert or remove. A successful search returns the index with an OK status
+while a failed search indicates a failure error. */
+static ccc_ucount
+find_key(struct ccc_fhmap_ const *const h, void const *const key,
+         uint64_t const hash)
 {
-    *hash_at(h, ccc_buf_i(&h->buf_, e).count) = CCC_FHM_EMPTY;
-    size_t const cap = h->buf_.capacity_;
-    size_t const elem_sz = h->buf_.elem_sz_;
-    void *const tmp = ccc_buf_at(&h->buf_, 0);
-    size_t i = ccc_buf_i(&h->buf_, e).count;
-    size_t next = increment(cap, i);
+    ccc_fhm_tag const tag = to_tag(hash);
+    size_t const mask = h->mask_;
+    struct triangular_seq seq = {.i = hash & mask, .stride = 0};
     do
     {
-        void *const next_slot = ccc_buf_at(&h->buf_, next);
-        uint64_t const next_hash = elem_in_slot(h, next_slot)->hash_;
-        if (next_hash == CCC_FHM_EMPTY
-            || !distance(cap, next, to_i(cap, next_hash)))
+        group const g = load_group(&h->tag_[seq.i]);
+        index_mask m = match_tag(g, tag);
+        for (size_t i_match = lowest_on_index(m); i_match != CCC_FHM_GROUP_SIZE;
+             i_match = next_index(&m))
         {
-            elem_in_slot(h, tmp)->hash_ = CCC_FHM_EMPTY;
-            (void)ccc_buf_size_minus(&h->buf_, 1);
-            return;
+            i_match = (seq.i + i_match) & mask;
+            if (h->eq_fn_((ccc_key_cmp){.key_lhs = key,
+                                        .user_type_rhs = data_at(h, i_match),
+                                        .aux = h->aux_}))
+            {
+                return (ccc_ucount){.count = i_match};
+            }
         }
-        swap(tmp, next_slot, ccc_buf_at(&h->buf_, i), elem_sz);
-        i = next;
-        next = increment(cap, next);
+        if (is_index_on(match_empty(g)))
+        {
+            return (ccc_ucount){.error = CCC_RESULT_FAIL};
+        }
+        seq.stride += CCC_FHM_GROUP_SIZE;
+        seq.i += seq.stride;
+        seq.i &= mask;
+    } while (1);
+}
+
+/* Finds an insert slot or loops forever. The caller of this function must know
+that there is an available empty or deleted slot in the table. */
+static size_t
+find_known_insert_slot(struct ccc_fhmap_ const *const h, uint64_t const hash)
+{
+    size_t const mask = h->mask_;
+    struct triangular_seq seq = {.i = hash & mask, .stride = 0};
+    do
+    {
+        size_t const i = lowest_on_index(
+            match_empty_or_deleted(load_group(&h->tag_[seq.i])));
+        if (likely(i != CCC_FHM_GROUP_SIZE))
+        {
+            return i;
+        }
+        seq.stride += CCC_FHM_GROUP_SIZE;
+        seq.i += seq.stride;
+        seq.i &= mask;
     } while (1);
 }
 
 static ccc_result
-maybe_resize(struct ccc_fhmap_ *const h)
+maybe_rehash(struct ccc_fhmap_ *const h, size_t const to_add)
 {
-    if (unlikely(h->buf_.capacity_ && h->buf_.sz_ < NUM_SWAP_SLOTS))
-    {
-        (void)memset(h->buf_.mem_, CCC_FHM_EMPTY,
-                     h->buf_.capacity_ * h->buf_.elem_sz_);
-        (void)ccc_buf_size_set(&h->buf_, NUM_SWAP_SLOTS);
-    }
-    if (h->buf_.capacity_
-        && (double)(h->buf_.sz_) / (double)h->buf_.capacity_ <= load_factor)
-    {
-        return CCC_RESULT_OK;
-    }
-    if (!h->buf_.alloc_)
+    if (unlikely(!h->mask_ && !h->alloc_fn_))
     {
         return CCC_RESULT_NO_ALLOC;
     }
-    struct ccc_fhmap_ new_hash = *h;
-    new_hash.buf_.sz_ = 0;
-    new_hash.buf_.capacity_
-        = new_hash.buf_.capacity_ ? next_prime(h->buf_.sz_ * 2) : primes[0];
-    if (new_hash.buf_.capacity_ <= h->buf_.capacity_)
+    size_t required_cap = ((h->sz_ + to_add) * 8) / 7;
+    if ((required_cap & (required_cap - 1)) != 0)
     {
-        return CCC_RESULT_MEM_ERROR;
+        required_cap = next_power_of_two(required_cap);
     }
-    new_hash.buf_.mem_ = new_hash.buf_.alloc_(
-        NULL, h->buf_.elem_sz_ * new_hash.buf_.capacity_, h->buf_.aux_);
-    if (!new_hash.buf_.mem_)
+    if (unlikely(!h->init_))
     {
-        return CCC_RESULT_MEM_ERROR;
-    }
-    /* Empty is intentionally chosen as zero so every byte is just set to
-       0 in this new array. */
-    (void)memset(ccc_buf_begin(&new_hash.buf_), CCC_FHM_EMPTY,
-                 new_hash.buf_.capacity_ * new_hash.buf_.elem_sz_);
-    (void)ccc_buf_size_set(&new_hash.buf_, NUM_SWAP_SLOTS);
-    for (void *slot = ccc_buf_begin(&h->buf_);
-         slot != ccc_buf_capacity_end(&h->buf_);
-         slot = ccc_buf_next(&h->buf_, slot))
-    {
-        struct ccc_fhmap_elem_ const *const e = elem_in_slot(h, slot);
-        if (e->hash_ != CCC_FHM_EMPTY)
+        if (h->mask_)
         {
-            struct ccc_ent_ const new_ent
-                = find(&new_hash, key_in_slot(h, slot), e->hash_);
-            insert(&new_hash, slot, e->hash_,
-                   ccc_buf_i(&new_hash.buf_, new_ent.e_).count);
+            if (!h->tag_ || h->mask_ + 1 < required_cap)
+            {
+                return CCC_RESULT_MEM_ERROR;
+            }
+            if (h->mask_ + 1 < CCC_FHM_GROUP_SIZE
+                || ((h->mask_ + 1) & h->mask_) != 0)
+            {
+                return CCC_RESULT_ARG_ERROR;
+            }
+            (void)memset(h->tag_, CCC_FHM_EMPTY,
+                         (h->mask_ + 1) + CCC_FHM_GROUP_SIZE);
         }
+        h->init_ = CCC_TRUE;
     }
-    if (ccc_buf_alloc(&h->buf_, 0, h->buf_.alloc_) != CCC_RESULT_OK)
+    if (unlikely(!h->mask_))
     {
-        *h = new_hash;
-        return CCC_RESULT_MEM_ERROR;
+        size_t const total_bytes = ((required_cap + 1) * h->elem_sz_)
+                                   + (required_cap + CCC_FHM_GROUP_SIZE);
+        void *const buf = h->alloc_fn_(NULL, total_bytes, h->aux_);
+        if (!buf)
+        {
+            return CCC_RESULT_MEM_ERROR;
+        }
+        h->mask_ = required_cap - 1;
+        h->data_ = buf;
+        h->avail_ = (required_cap / 8) * 7;
+        h->tag_ = (ccc_fhm_tag *)(((char *)buf + total_bytes)
+                                  - (required_cap + CCC_FHM_GROUP_SIZE));
+        (void)memset(h->tag_, CCC_FHM_EMPTY, required_cap + CCC_FHM_GROUP_SIZE);
     }
-    *h = new_hash;
+    if (likely(h->avail_))
+    {
+        return CCC_RESULT_OK;
+    }
+    size_t const current_cap = ((h->mask_ + 1) / 8) * 7;
+    if (h->alloc_fn_ && (h->sz_ + to_add) > current_cap / 2)
+    {
+        assert(h->alloc_fn_);
+        return rehash_resize(h, to_add);
+    }
+    rehash_in_place(h);
     return CCC_RESULT_OK;
 }
 
-static size_t
-next_prime(size_t const n)
+static void
+rehash_in_place(struct ccc_fhmap_ *const h)
 {
-    /* Compiler should help us out here. No fancy binary search/tricks to
-       tell how far along the list we are. The table is small already. */
-    for (size_t i = 0; i < PRIMES_SIZE; ++i)
+    assert((h->mask_ + 1) % CCC_FHM_GROUP_SIZE == 0);
+    size_t const mask = h->mask_;
+    size_t const allowed_cap = ((mask + 1) / 8) * 7;
+    for (size_t i = 0; i < mask + 1; i += CCC_FHM_GROUP_SIZE)
     {
-        if (primes[i] > n)
+        store_group(&h->tag_[i], make_deleted_empty_and_full_deleted(
+                                     load_group(&h->tag_[i])));
+    }
+    (void)memcpy(h->tag_ + (mask + 1), h->tag_, CCC_FHM_GROUP_SIZE);
+    for (size_t i = 0; i < mask + 1; ++i)
+    {
+        if (h->tag_[i].v == CCC_FHM_DELETED)
         {
-            return primes[i];
+            set_tag(h, (ccc_fhm_tag){CCC_FHM_EMPTY}, i);
+            --h->sz_;
         }
     }
-    return 0;
+    h->avail_ = allowed_cap - h->sz_;
+    for (size_t i = 0; i < mask + 1; ++i)
+    {
+        if (h->tag_[i].v != CCC_FHM_DELETED)
+        {
+            continue;
+        }
+        do
+        {
+            uint64_t const hash = h->hash_fn_(
+                (ccc_user_key){.user_key = key_at(h, i), .aux = h->aux_});
+            ccc_fhm_tag const hash_tag = to_tag(hash);
+            size_t const new_slot = find_known_insert_slot(h, hash);
+            size_t const hash_pos = (hash & mask);
+            size_t const group_a = ((i - hash_pos) & mask) / CCC_FHM_GROUP_SIZE;
+            size_t const group_b
+                = ((new_slot - hash_pos) & mask) / CCC_FHM_GROUP_SIZE;
+            if (group_a == group_b)
+            {
+                set_tag(h, hash_tag, i);
+                break; /* continues outer loop */
+            }
+            ccc_fhm_tag const prev = h->tag_[new_slot];
+            set_tag(h, hash_tag, new_slot);
+            if (prev.v == CCC_FHM_EMPTY)
+            {
+                set_tag(h, (ccc_fhm_tag){CCC_FHM_EMPTY}, i);
+                (void)memcpy(data_at(h, new_slot), data_at(h, i), h->elem_sz_);
+                break; /* continues outer loop */
+            }
+            assert(prev.v == CCC_FHM_DELETED);
+            swap(h->data_, data_at(h, i), data_at(h, new_slot), h->elem_sz_);
+        } while (1);
+    }
+    h->avail_ = allowed_cap - h->sz_;
 }
 
-static inline uint64_t *
-hash_at(struct ccc_fhmap_ const *const h, size_t const i)
+static ccc_result
+rehash_resize(struct ccc_fhmap_ *const h, size_t const to_add)
 {
-    return &((struct ccc_fhmap_elem_ *)((char *)ccc_buf_at(&h->buf_, i)
-                                        + h->hash_elem_offset_))
-                ->hash_;
+    assert(((h->mask_ + 1) & h->mask_) == 0);
+    size_t const new_pow2_cap = next_power_of_two(h->mask_ + 1 + to_add) << 1;
+    if (new_pow2_cap < (h->mask_ + 1))
+    {
+        return CCC_RESULT_MEM_ERROR;
+    }
+    size_t const prev_bytes
+        = ((h->mask_ + 2) * h->elem_sz_) + (h->mask_ + 1 + CCC_FHM_GROUP_SIZE);
+    size_t const total_bytes = ((new_pow2_cap + 1) * h->elem_sz_)
+                               + (new_pow2_cap + CCC_FHM_GROUP_SIZE);
+    if (total_bytes < prev_bytes)
+    {
+        return CCC_RESULT_MEM_ERROR;
+    }
+    void *const new_buf = h->alloc_fn_(NULL, total_bytes, h->aux_);
+    if (!new_buf)
+    {
+        return CCC_RESULT_MEM_ERROR;
+    }
+    struct ccc_fhmap_ new_h = *h;
+    new_h.sz_ = 0;
+    new_h.avail_ = (new_pow2_cap / 8) * 7;
+    new_h.mask_ = new_pow2_cap - 1;
+    new_h.tag_ = (ccc_fhm_tag *)(((char *)new_buf + total_bytes) - new_pow2_cap
+                                 - CCC_FHM_GROUP_SIZE);
+    new_h.data_ = new_buf;
+    (void)memset(new_h.tag_, CCC_FHM_EMPTY, new_pow2_cap + CCC_FHM_GROUP_SIZE);
+    for (size_t i = 0; i < (h->mask_ + 1); ++i)
+    {
+        if (is_full(h->tag_[i]))
+        {
+            uint64_t const hash = h->hash_fn_(
+                (ccc_user_key){.user_key = key_at(h, i), .aux = h->aux_});
+            size_t const new_i = find_known_insert_slot(&new_h, hash);
+            set_tag(&new_h, to_tag(hash), new_i);
+            (void)memcpy(data_at(&new_h, new_i), data_at(h, i), new_h.elem_sz_);
+        }
+    }
+    new_h.avail_ -= h->sz_;
+    new_h.sz_ = h->sz_;
+    (void)h->alloc_fn_(h->data_, 0, h->aux_);
+    *h = new_h;
+    return CCC_RESULT_OK;
 }
 
-static inline uint64_t
-filter(struct ccc_fhmap_ const *const h, void const *const key)
+static inline void
+set_tag(struct ccc_fhmap_ *const h, ccc_fhm_tag const m, size_t const i)
 {
-    uint64_t const hash
-        = h->hash_fn_((ccc_user_key){.user_key = key, .aux = h->buf_.aux_});
-    return hash == CCC_FHM_EMPTY ? hash + 1 : hash;
+    size_t const replica_byte
+        = ((i - CCC_FHM_GROUP_SIZE) & h->mask_) + CCC_FHM_GROUP_SIZE;
+    h->tag_[i] = m;
+    h->tag_[replica_byte] = m;
 }
 
-static inline struct ccc_fhmap_elem_ *
-elem_in_slot(struct ccc_fhmap_ const *const h, void const *const slot)
+static inline ccc_tribool
+is_full(ccc_fhm_tag const m)
 {
-    return (struct ccc_fhmap_elem_ *)((char *)slot + h->hash_elem_offset_);
+    return (m.v & CCC_FHM_DELETED) == 0;
+}
+
+static inline ccc_tribool
+is_constant(ccc_fhm_tag const m)
+{
+    return (m.v & TAG_MSB) != 0;
+}
+
+static inline ccc_tribool
+is_empty_constant(ccc_fhm_tag const m)
+{
+    assert(is_constant(m));
+    return (m.v & TAG_LSB) != 0;
+}
+
+static inline ccc_fhm_tag
+to_tag(uint64_t const hash)
+{
+    return (ccc_fhm_tag){(hash >> ((sizeof(hash) * CHAR_BIT) - 7))
+                         & LOWER_7_BITS_MASK};
 }
 
 static inline void *
-key_in_slot(struct ccc_fhmap_ const *const h, void const *const slot)
+key_at(struct ccc_fhmap_ const *const h, size_t const i)
 {
-    return (char *)slot + h->key_offset_;
-}
-
-static inline size_t
-distance(size_t const capacity, size_t const i, size_t const j)
-{
-    return i < j ? (capacity - j) + i - NUM_SWAP_SLOTS : i - j;
-}
-
-static inline size_t
-increment(size_t const capacity, size_t const i)
-{
-    return (i + 1) >= capacity ? LAST_SWAP_SLOT + 1 : i + 1;
-}
-
-static inline size_t
-decrement(size_t const capacity, size_t const i)
-{
-    return i <= NUM_SWAP_SLOTS ? capacity - 1 : i - 1;
+    assert(i <= h->mask_);
+    return (char *)(h->tag_ - ((i + 1) * h->elem_sz_)) + h->key_offset_;
 }
 
 static inline void *
-struct_base(struct ccc_fhmap_ const *const h,
-            struct ccc_fhmap_elem_ const *const e)
+data_at(struct ccc_fhmap_ const *const h, size_t const i)
 {
-    return ((char *)&e->hash_) - h->hash_elem_offset_;
+    assert(i <= h->mask_);
+    return h->tag_ - ((i + 1) * h->elem_sz_);
 }
 
 static inline void
 swap(char tmp[const], void *const a, void *const b, size_t const ab_size)
 {
-    if (a == b || !a || !b)
+    if (unlikely(!a || !b || a == b))
     {
         return;
     }
@@ -1007,247 +590,382 @@ swap(char tmp[const], void *const a, void *const b, size_t const ab_size)
     (void)memcpy(b, tmp, ab_size);
 }
 
-/** This function converts a hash to its home index in the table. Because this
-operation is in the hot path of all table functions and Robin Hood relies on
-home slot distance calculations, this function tries to use Daniel Lemire's
-fastrange library modulo calculation alternative:
-    https://github.com/lemire/fastrange
-Such a technique is helpful because the prime table capacities used to mitigate
-primary clustering of Robin Hood hash tables make efficient modulo calculations
-more difficult than power of two table capacities. The appropriate license for
-this technique is included immediately following this function and only applies
-to this function in this source file. */
 static inline size_t
-to_i(size_t const capacity, uint64_t hash)
+next_power_of_two(size_t const n)
 {
-#ifdef __SIZEOF_INT128__
-    hash = (uint64_t)(((__uint128_t)hash * (__uint128_t)capacity) >> 64);
-#else
-    hash %= capacity;
-#endif
-    return (size_t)hash <= LAST_SWAP_SLOT ? LAST_SWAP_SLOT + 1 : (size_t)hash;
+    return n <= 1 ? n : (SIZE_MAX >> countl_0_size_t(n - 1)) + 1;
 }
 
-/** The following Apache License applies only to the above function. This
-function uses one line from Daniel Lemire's fastrange library to compute the
-modulo of a number range with multiplication and shifts rather than  modulo
-division. Specifically, the 128 bit widening is from the fastrange library which
-can be found here: https://github.com/lemire/fastrange
+/*=====================   Intrinsics and Generics   =========================*/
 
-Below is the required license for the code in the above function. Lemire's
-copyright applies to the integer widening line of code as he copyrights the
-source file from which it came with his name and the year. The notice in
-https://github.com/lemire/fastrange/fastrange.h appears as follows:
+/** Below are the implementations of the SIMD or bitwise operations needed to
+run a search on multiple entries in the hash table simultaneously. For now,
+the only container that will use these operations is this one so there is no
+need to break out different headers and sources and clutter the src directory.
+x86 is the only platform that gets the full benefit of SIMD. Apple and all
+other platforms will get a portable implementation due to concerns over NEON
+speed of vectorized instructions. However, loading up groups into a uint64_t is
+still good and counts simultaneous operations just not the type that uses CPU
+vector lanes for a single instruction. */
 
-```
-Fair maps to intervals without division.
-Reference :
-http://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
+/* We can vectorize for x86 only. */
+#if defined(__x86_64) && defined(__SSE2__) && !defined(CCC_FHM_PORTABLE)
 
-(c) Daniel Lemire
-Apache License 2.0
-```
+/*========================  Index Mask Implementations   ====================*/
 
-The license for the cited repository is included below for completeness.
+static inline ccc_tribool
+is_index_on(index_mask const m)
+{
+    return m.v != 0;
+}
 
-                                 Apache License
-                           Version 2.0, January 2004
-                        http://www.apache.org/licenses/
+static inline size_t
+lowest_on_index(index_mask const m)
+{
+    return countr_0(m);
+}
 
-   TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
+static inline size_t
+trailing_zeros(index_mask const m)
+{
+    return countr_0(m);
+}
 
-   1. Definitions.
+static inline size_t
+leading_zeros(index_mask const m)
+{
+    return countl_0(m);
+}
 
-      "License" shall mean the terms and conditions for use, reproduction,
-      and distribution as defined by Sections 1 through 9 of this document.
+static inline size_t
+next_index(index_mask *const m)
+{
+    assert(m);
+    size_t const index = lowest_on_index(*m);
+    m->v &= (m->v - 1);
+    return index;
+}
 
-      "Licensor" shall mean the copyright owner or entity authorized by
-      the copyright owner that is granting the License.
+/*=========================  Group Implementations   ========================*/
 
-      "Legal Entity" shall mean the union of the acting entity and all
-      other entities that control, are controlled by, or are under common
-      control with that entity. For the purposes of this definition,
-      "control" means (i) the power, direct or indirect, to cause the
-      direction or management of such entity, whether by contract or
-      otherwise, or (ii) ownership of fifty percent (50%) or more of the
-      outstanding shares, or (iii) beneficial ownership of such entity.
+static inline group
+load_group(ccc_fhm_tag const *const src)
+{
+    return (group){_mm_load_si128((__m128i *)&src->v)};
+}
 
-      "You" (or "Your") shall mean an individual or Legal Entity
-      exercising permissions granted by this License.
+static inline void
+store_group(ccc_fhm_tag *const dst, group const src)
+{
+    _mm_store_si128((__m128i *)&dst->v, src.v);
+}
 
-      "Source" form shall mean the preferred form for making modifications,
-      including but not limited to software source code, documentation
-      source, and configuration files.
+static inline index_mask
+match_tag(group const g, ccc_fhm_tag const m)
+{
+    return (index_mask){
+        _mm_movemask_epi8(_mm_cmpeq_epi8(g.v, _mm_set1_epi8((int8_t)m.v)))};
+}
 
-      "Object" form shall mean any form resulting from mechanical
-      transformation or translation of a Source form, including but
-      not limited to compiled object code, generated documentation,
-      and conversions to other media types.
+static inline index_mask
+match_empty(group const g)
+{
+    return match_tag(g, (ccc_fhm_tag){CCC_FHM_EMPTY});
+}
 
-      "Work" shall mean the work of authorship, whether in Source or
-      Object form, made available under the License, as indicated by a
-      copyright notice that is included in or attached to the work
-      (an example is provided in the Appendix below).
+static inline index_mask
+match_empty_or_deleted(group const g)
+{
+    static_assert(sizeof(int) >= sizeof(uint16_t));
+    return (index_mask){_mm_movemask_epi8(g.v)};
+}
 
-      "Derivative Works" shall mean any work, whether in Source or Object
-      form, that is based on (or derived from) the Work and for which the
-      editorial revisions, annotations, elaborations, or other modifications
-      represent, as a whole, an original work of authorship. For the purposes
-      of this License, Derivative Works shall not include works that remain
-      separable from, or merely link (or bind by name) to the interfaces of,
-      the Work and Derivative Works thereof.
+static inline index_mask
+match_full(group const g)
+{
+    index_mask m = match_empty_or_deleted(g);
+    m.v = ~m.v;
+    return m;
+}
 
-      "Contribution" shall mean any work of authorship, including
-      the original version of the Work and any modifications or additions
-      to that Work or Derivative Works thereof, that is intentionally
-      submitted to Licensor for inclusion in the Work by the copyright owner
-      or by an individual or Legal Entity authorized to submit on behalf of
-      the copyright owner. For the purposes of this definition, "submitted"
-      means any form of electronic, verbal, or written communication sent
-      to the Licensor or its representatives, including but not limited to
-      communication on electronic mailing lists, source code control systems,
-      and issue tracking systems that are managed by, or on behalf of, the
-      Licensor for the purpose of discussing and improving the Work, but
-      excluding communication that is conspicuously marked or otherwise
-      designated in writing by the copyright owner as "Not a Contribution."
+static inline group
+make_deleted_empty_and_full_deleted(group const g)
+{
+    __m128i const zero = _mm_setzero_si128();
+    __m128i const match_constants = _mm_cmpgt_epi8(zero, g.v);
+    return (group){
+        _mm_or_si128(match_constants, _mm_set1_epi8((int8_t)CCC_FHM_DELETED))};
+}
 
-      "Contributor" shall mean Licensor and any individual or Legal Entity
-      on behalf of whom a Contribution has been received by Licensor and
-      subsequently incorporated within the Work.
+/*====================  Bit Counting for Index Mask   =======================*/
 
-   2. Grant of Copyright License. Subject to the terms and conditions of
-      this License, each Contributor hereby grants to You a perpetual,
-      worldwide, non-exclusive, no-charge, royalty-free, irrevocable
-      copyright license to reproduce, prepare Derivative Works of,
-      publicly display, publicly perform, sublicense, and distribute the
-      Work and such Derivative Works in Source or Object form.
+#    if defined(__has_builtin) && __has_builtin(__builtin_ctz)                 \
+        && __has_builtin(__builtin_clz) && __has_builtin(__builtin_clzl)
 
-   3. Grant of Patent License. Subject to the terms and conditions of
-      this License, each Contributor hereby grants to You a perpetual,
-      worldwide, non-exclusive, no-charge, royalty-free, irrevocable
-      (except as stated in this section) patent license to make, have made,
-      use, offer to sell, sell, import, and otherwise transfer the Work,
-      where such license applies only to those patent claims licensable
-      by such Contributor that are necessarily infringed by their
-      Contribution(s) alone or by combination of their Contribution(s)
-      with the Work to which such Contribution(s) was submitted. If You
-      institute patent litigation against any entity (including a
-      cross-claim or counterclaim in a lawsuit) alleging that the Work
-      or a Contribution incorporated within the Work constitutes direct
-      or contributory patent infringement, then any patent licenses
-      granted to You under this License for that Work shall terminate
-      as of the date such litigation is filed.
+static_assert(sizeof((index_mask){}.v) <= sizeof(unsigned));
 
-   4. Redistribution. You may reproduce and distribute copies of the
-      Work or Derivative Works thereof in any medium, with or without
-      modifications, and in Source or Object form, provided that You
-      meet the following conditions:
+static inline unsigned
+countr_0(index_mask const m)
+{
+    return m.v ? __builtin_ctz(m.v) : CCC_FHM_GROUP_SIZE;
+}
 
-      (a) You must give any other recipients of the Work or
-          Derivative Works a copy of this License; and
+static inline unsigned
+countl_0(index_mask const m)
+{
+    return m.v ? __builtin_clz(m.v << CCC_FHM_GROUP_SIZE) : CCC_FHM_GROUP_SIZE;
+}
 
-      (b) You must cause any modified files to carry prominent notices
-          stating that You changed the files; and
+static inline unsigned
+countl_0_size_t(size_t const n)
+{
+    static_assert(sizeof(size_t) == sizeof(unsigned long));
+    return n ? __builtin_clzl(n) : sizeof(size_t) * CHAR_BIT;
+}
 
-      (c) You must retain, in the Source form of any Derivative Works
-          that You distribute, all copyright, patent, trademark, and
-          attribution notices from the Source form of the Work,
-          excluding those notices that do not pertain to any part of
-          the Derivative Works; and
+#    else /* !defined(__has_builtin) || !__has_builtin(__builtin_ctz)          \
+        || !__has_builtin(__builtin_clz) || !__has_builtin(__builtin_clzl) */
 
-      (d) If the Work includes a "NOTICE" text file as part of its
-          distribution, then any Derivative Works that You distribute must
-          include a readable copy of the attribution notices contained
-          within such NOTICE file, excluding those notices that do not
-          pertain to any part of the Derivative Works, in at least one
-          of the following places: within a NOTICE text file distributed
-          as part of the Derivative Works; within the Source form or
-          documentation, if provided along with the Derivative Works; or,
-          within a display generated by the Derivative Works, if and
-          wherever such third-party notices normally appear. The contents
-          of the NOTICE file are for informational purposes only and
-          do not modify the License. You may add Your own attribution
-          notices within Derivative Works that You distribute, alongside
-          or as an addendum to the NOTICE text from the Work, provided
-          that such additional attribution notices cannot be construed
-          as modifying the License.
+static inline unsigned
+countr_0(index_mask m)
+{
+    if (!m.v)
+    {
+        return CCC_FHM_GROUP_SIZE;
+    }
+    unsigned cnt = 0;
+    for (; m.v; cnt += !!(m.v & 1U), m.v >>= 1U)
+    {}
+    return cnt;
+}
 
-      You may add Your own copyright statement to Your modifications and
-      may provide additional or different license terms and conditions
-      for use, reproduction, or distribution of Your modifications, or
-      for any such Derivative Works as a whole, provided Your use,
-      reproduction, and distribution of the Work otherwise complies with
-      the conditions stated in this License.
+static inline unsigned
+countl_0(index_mask m)
+{
+    if (!m.v)
+    {
+        return CCC_FHM_GROUP_SIZE;
+    }
+    m.v <<= CCC_FHM_GROUP_SIZE;
+    unsigned cnt = 0;
+    for (; !(m.v & INDEX_MASK_MSB); ++cnt, m.v <<= 1U)
+    {}
+    return cnt;
+}
 
-   5. Submission of Contributions. Unless You explicitly state otherwise,
-      any Contribution intentionally submitted for inclusion in the Work
-      by You to the Licensor shall be under the terms and conditions of
-      this License, without any additional terms or conditions.
-      Notwithstanding the above, nothing herein shall supersede or modify
-      the terms of any separate license agreement you may have executed
-      with Licensor regarding such Contributions.
+static inline unsigned
+countl_0_size_t(size_t n)
+{
+    if (!n)
+    {
+        return sizeof(size_t) * CHAR_BIT;
+    }
+    unsigned cnt = 0;
+    for (; !(n & SIZE_T_MSB); ++cnt, n <<= 1U)
+    {}
+    return cnt;
+}
 
-   6. Trademarks. This License does not grant permission to use the trade
-      names, trademarks, service marks, or product names of the Licensor,
-      except as required for reasonable and customary use in describing the
-      origin of the Work and reproducing the content of the NOTICE file.
+#    endif /* defined(__has_builtin) && __has_builtin(__builtin_ctz)           \
+        && __has_builtin(__builtin_clz) && __has_builtin(__builtin_clzl) */
 
-   7. Disclaimer of Warranty. Unless required by applicable law or
-      agreed to in writing, Licensor provides the Work (and each
-      Contributor provides its Contributions) on an "AS IS" BASIS,
-      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-      implied, including, without limitation, any warranties or conditions
-      of TITLE, NON-INFRINGEMENT, MERCHANTABILITY, or FITNESS FOR A
-      PARTICULAR PURPOSE. You are solely responsible for determining the
-      appropriateness of using or redistributing the Work and assume any
-      risks associated with Your exercise of permissions under this License.
+#else /* !defined(__x86_64) || !defined(__SSE2__) ||                           \
+         !defined(CCC_FHM_PORTABLE) */
 
-   8. Limitation of Liability. In no event and under no legal theory,
-      whether in tort (including negligence), contract, or otherwise,
-      unless required by applicable law (such as deliberate and grossly
-      negligent acts) or agreed to in writing, shall any Contributor be
-      liable to You for damages, including any direct, indirect, special,
-      incidental, or consequential damages of any character arising as a
-      result of this License or out of the use or inability to use the
-      Work (including but not limited to damages for loss of goodwill,
-      work stoppage, computer failure or malfunction, or any and all
-      other commercial damages or losses), even if such Contributor
-      has been advised of the possibility of such damages.
+/* What follows is the generic portable implementation when high width SIMD
+can't be achieved. For now this means NEON on Apple defaults to generic but
+this will likely change in the future as NEON improves. */
 
-   9. Accepting Warranty or Additional Liability. While redistributing
-      the Work or Derivative Works thereof, You may choose to offer,
-      and charge a fee for, acceptance of support, warranty, indemnity,
-      or other liability obligations and/or rights consistent with this
-      License. However, in accepting such obligations, You may act only
-      on Your own behalf and on Your sole responsibility, not on behalf
-      of any other Contributor, and only if You agree to indemnify,
-      defend, and hold each Contributor harmless for any liability
-      incurred by, or claims asserted against, such Contributor by reason
-      of your accepting any such warranty or additional liability.
+/*==========================    Bit Helpers    ==============================*/
 
-   END OF TERMS AND CONDITIONS
+#    if defined(__has_builtin) && __has_builtin(__builtin_ctzl)                \
+        && __has_builtin(__builtin_clzl)
 
-   APPENDIX: How to apply the Apache License to your work.
+static_assert(sizeof((index_mask){}.v) == sizeof(long));
 
-      To apply the Apache License to your work, attach the following
-      boilerplate notice, with the fields enclosed by brackets "{}"
-      replaced with your own identifying information. (Don't include
-      the brackets!)  The text should be enclosed in the appropriate
-      comment syntax for the file format. We also recommend that a
-      file or class name and description of purpose be included on the
-      same "printed page" as the copyright notice for easier
-      identification within third-party archives.
+static inline unsigned
+countr_0(index_mask const m)
+{
+    return m.v ? __builtin_ctzl(m.v) / CCC_FHM_GROUP_SIZE : CCC_FHM_GROUP_SIZE;
+}
 
-   Copyright {yyyy} {name of copyright owner}
+static inline unsigned
+countl_0(index_mask const m)
+{
+    return m.v ? __builtin_clzl(m.v) / CCC_FHM_GROUP_SIZE : CCC_FHM_GROUP_SIZE;
+}
 
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
+static inline unsigned
+countl_0_size_t(size_t const n)
+{
+    static_assert(sizeof(size_t) == sizeof(unsigned long));
+    return n ? __builtin_clzl(n) : sizeof(size_t) * CHAR_BIT;
+}
 
-       http://www.apache.org/licenses/LICENSE-2.0
+#    else /* defined(__has_builtin) && __has_builtin(__builtin_ctzl) &&        \
+             __has_builtin(__builtin_clzl) */
 
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License. */
+static inline unsigned
+countr_0(index_mask m)
+{
+    if (!m.v)
+    {
+        return CCC_FHM_GROUP_SIZE;
+    }
+    unsigned cnt = 0;
+    for (; m.v; cnt += !!(m.v & 1U), m.v >>= 1U)
+    {}
+    return cnt / CCC_FHM_GROUP_SIZE;
+}
+
+static inline unsigned
+countl_0(index_mask m)
+{
+    if (!m.v)
+    {
+        return CCC_FHM_GROUP_SIZE;
+    }
+    unsigned cnt = 0;
+    for (; !(m.v & INDEX_MASK_MSB); ++cnt, m.v <<= 1U)
+    {}
+    return cnt / CCC_FHM_GROUP_SIZE;
+}
+
+static inline unsigned
+countl_0_size_t(size_t n)
+{
+    if (!n)
+    {
+        return sizeof(size_t) * CHAR_BIT;
+    }
+    unsigned cnt = 0;
+    for (; !(n & SIZE_T_MSB); ++cnt, n <<= 1U)
+    {}
+    return cnt;
+}
+
+#    endif /* !defined(__has_builtin) || !__has_builtin(__builtin_ctzl) ||     \
+              !__has_builtin(__builtin_clzl) */
+
+/* Returns 1 aka true if platform is little endian otherwise false for big
+endian. */
+static inline int
+is_little_endian(void)
+{
+    unsigned int x = 1;
+    char *c = (char *)&x;
+    return (int)*c;
+}
+
+static inline index_mask
+to_little_endian(index_mask m)
+{
+    if (is_little_endian())
+    {
+        return m;
+    }
+#    if defined(__has_builtin) && __has_builtin(__builtin_bswap64)
+    m.v = __builtin_bswap64(m.v);
+#    else
+    m.v = (m.v & 0x00000000FFFFFFFF) << 32 | (m.v & 0xFFFFFFFF00000000) >> 32;
+    m.v = (m.v & 0x0000FFFF0000FFFF) << 16 | (m.v & 0xFFFF0000FFFF0000) >> 16;
+    m.v = (m.v & 0x00FF00FF00FF00FF) << 8 | (m.v & 0xFF00FF00FF00FF00) >> 8;
+#    endif
+    return m;
+}
+
+/*========================  Index Mask Implementations   ====================*/
+
+static inline ccc_tribool
+is_index_on(index_mask const m)
+{
+    return m.v != 0;
+}
+
+static inline size_t
+lowest_on_index(index_mask const m)
+{
+    return countr_0(m);
+}
+
+static inline size_t
+trailing_zeros(index_mask const m)
+{
+    return countr_0(m);
+}
+
+static inline size_t
+leading_zeros(index_mask const m)
+{
+    return countl_0(m);
+}
+
+static inline size_t
+next_index(index_mask *const m)
+{
+    assert(m);
+    size_t const index = lowest_on_index(*m);
+    m->v &= (m->v - 1);
+    return index;
+}
+
+/*=========================  Group Implementations   ========================*/
+
+static inline group
+load_group(ccc_fhm_tag const *const src)
+{
+    group g;
+    (void)memcpy(&g, src, sizeof(g));
+    return g;
+}
+
+static inline void
+store_group(ccc_fhm_tag *const dst, group const src)
+{
+    (void)memcpy(dst, &src, sizeof(src));
+}
+
+static inline index_mask
+match_tag(group g, ccc_fhm_tag const m)
+{
+    group const cmp = {g.v
+                       ^ ((((typeof(g.v))m.v) << (TAG_BITS * 7UL))
+                          | (((typeof(g.v))m.v) << (TAG_BITS * 6UL))
+                          | (((typeof(g.v))m.v) << (TAG_BITS * 5UL))
+                          | (((typeof(g.v))m.v) << (TAG_BITS * 4UL))
+                          | (((typeof(g.v))m.v) << (TAG_BITS * 3UL))
+                          | (((typeof(g.v))m.v) << (TAG_BITS * 2UL))
+                          | (((typeof(g.v))m.v) << TAG_BITS) | m.v)};
+    return to_little_endian(
+        (index_mask){(cmp.v - GROUP_LSB) & ~cmp.v & GROUP_DELETED});
+}
+
+static inline index_mask
+match_empty(group const g)
+{
+    return to_little_endian((index_mask){g.v & (g.v << 1) & GROUP_DELETED});
+}
+
+static inline index_mask
+match_empty_or_deleted(group const g)
+{
+    return to_little_endian((index_mask){g.v & GROUP_DELETED});
+}
+
+static inline index_mask
+match_full(group const g)
+{
+    index_mask res = match_empty_or_deleted(g);
+    res.v = ~res.v;
+    return res;
+}
+
+static inline group
+make_deleted_empty_and_full_deleted(group g)
+{
+    g.v = ~g.v & GROUP_DELETED;
+    g.v = ~g.v + (g.v >> (TAG_BITS - 1));
+    return g;
+}
+
+#endif /* defined(__x86_64) && defined(__SSE2__) */
